chore: move model_engine to models
