+++ beta_rec/models/cmn.py
@@ -4,7 +4,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from tqdm import tqdm
 
-from beta_rec.cores.torch_engine import Engine
+from beta_rec.models.torch_engine import ModelEngine
 from beta_rec.models.pairwise_gmf import truncated_normal_
 from beta_rec.models.vlml import VariableLengthMemoryLayer
 
@@ -129,7 +129,7 @@ class CollaborativeMemoryNetwork(nn.Module):
         return score
 
 
-class cmnEngine(Engine):
+class cmnEngine(ModelEngine):
     def __init__(self, config, user_embeddings, item_embeddings, item_user_list):
         self.config = config
         self.device = config["device_str"]
