+++ beta_rec/models/gmf.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 
-from beta_rec.cores.torch_engine import Engine
+from beta_rec.models.torch_engine import ModelEngine
 from beta_rec.utils.common_util import timeit
 
 
@@ -42,7 +42,7 @@ class GMF(torch.nn.Module):
         nn.init.normal_(self.embedding_user.weight, std=0.01)
 
 
-class GMFEngine(Engine):
+class GMFEngine(ModelEngine):
     """Engine for training & evaluating GMF model"""
 
     def __init__(self, config):
