+++ beta_rec/models/narm.py
@@ -10,7 +10,7 @@ from torch.optim.lr_scheduler import StepLR
 from torch.utils.data import DataLoader
 from tqdm import tqdm
 
-from beta_rec.cores.torch_engine import Engine
+from beta_rec.models.torch_engine import ModelEngine
 from beta_rec.datasets.seq_data_utils import SeqDataset, collate_fn
 
 
@@ -92,7 +92,7 @@ class NARM(nn.Module):
         ).to(self.device)
 
 
-class NARMEngine(Engine):
+class NARMEngine(ModelEngine):
     """Engine for training & evaluating NARM model"""
 
     def __init__(self, config):
