+++ beta_rec/models/ngcf.py
@@ -3,8 +3,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 import torch.sparse as sparse
 
-from beta_rec.cores.torch_engine import Engine
-
+from beta_rec.models.torch_engine import ModelEngine
 
 class NGCF(torch.nn.Module):
     """Model initialisation, embedding generation and prediction of NGCF
@@ -93,7 +92,7 @@ class NGCF(torch.nn.Module):
         return scores
 
 
-class NGCFEngine(Engine):
+class NGCFEngine(ModelEngine):
     # A class includes train an epoch and train a batch of NGCF
 
     def __init__(self, config):
