+++ beta_rec/models/pairwise_gmf.py
@@ -4,8 +4,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from tqdm import tqdm
 
-from beta_rec.cores.torch_engine import Engine
-
+from beta_rec.models.torch_engine import ModelEngine
 
 def truncated_normal_(tensor, mean=0, std=1):
     size = tensor.shape
@@ -62,7 +61,7 @@ class PairwiseGMF(nn.Module):
         pass
 
 
-class PairwiseGMFEngine(Engine):
+class PairwiseGMFEngine(ModelEngine):
     def __init__(self, config):
         self.config = config
         self.model = PairwiseGMF(config)
