+++ examples/train_mf2020.py
@@ -7,7 +7,7 @@ sys.path.append("../")
 from tqdm import tqdm
 
 from beta_rec.cores.train_engine import TrainEngine
-from beta_rec.datasets.nmf_data_utils import SampleGenerator
+from beta_rec.datasets.data_base import SampleGenerator
 from beta_rec.models.mf import MFEngine
 from beta_rec.utils.common_util import update_args
 from beta_rec.utils.monitor import Monitor
